requirements.txt
amqp==2.1.4
appnope==0.1.0
awscli==1.11.114
awsebcli==3.10.5
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
billiard==3.5.0.2
blessed==1.14.2
boto==2.43.0
botocore==1.5.92
cachetools==2.0.1
cement==2.8.2
certifi==2017.7.27.1
cffi==1.9.1
chardet==3.0.4
colorama==0.3.7
configparser==3.5.0
contextlib2==0.5.4
coverage==4.3.1
cryptography==1.7.2
decorator==4.0.10
dill==0.2.7.1
dj-database-url==0.4.1
Django==1.10.7
django-activeurl==0.1.10
django-allauth==0.28.0
django-anymail==0.10
django-appconf==1.0.2
django-autofixture==0.12.1
django-classy-tags==0.8.0
django-colorfield==0.1.10
django-compat==1.0.13
django-cors-headers==2.1.0
django-crispy-forms==1.6.1
django-debug-toolbar==1.6
django-filer==1.2.5
django-filter==1.0.4
django-flat-responsive==1.0.0
django-flat-theme==1.1.4
django-hijack==2.1.2
django-hijack-admin==2.1.2
django-materializecss-form==1.0.1
django-mptt==0.8.6
django-pkgconf==0.2.0
django-polymorphic==1.0.2
django-recaptcha==1.3.1
django-redis-cache==1.7.1
django-robots==2.0
django-storages-redux==1.3.2
django-versatileimagefield==1.6
django-webpack-loader==0.4.1
django-widget-tweaks==1.4.1
djangocms-admin-style==1.2.6.2
djangorestframework==3.5.4
djangorestframework-jwt==1.11.0
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
easy-thumbnails==2.4.1
enum34==1.1.6
Fabric==1.13.1
Faker==0.7.3
flake8==3.2.1
future==0.16.0
futures==3.1.1
gapic-google-cloud-datastore-v1==0.15.3
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
gapic-google-cloud-pubsub-v1==0.15.4
gapic-google-cloud-spanner-admin-database-v1==0.15.3
gapic-google-cloud-spanner-admin-instance-v1==0.15.3
gapic-google-cloud-spanner-v1==0.15.3
google-auth==1.0.2
google-cloud==0.27.0
google-cloud-bigquery==0.26.0
google-cloud-bigtable==0.26.0
google-cloud-core==0.26.0
google-cloud-datastore==1.2.0
google-cloud-dns==0.26.0
google-cloud-error-reporting==0.26.0
google-cloud-language==0.27.0
google-cloud-logging==1.2.0
google-cloud-monitoring==0.26.0
google-cloud-pubsub==0.27.0
google-cloud-resource-manager==0.26.0
google-cloud-runtimeconfig==0.26.0
google-cloud-spanner==0.26.0
google-cloud-speech==0.28.0
google-cloud-storage==1.3.2
google-cloud-translate==1.1.0
google-cloud-videointelligence==0.25.0
google-cloud-vision==0.26.0
google-gax==0.15.14
google-resumable-media==0.2.3
googleapis-common-protos==1.5.2
grpc-google-iam-v1==0.11.1
grpcio==1.4.0
gunicorn==19.6.0
httplib2==0.9.2
idna==2.6
ipaddress==1.0.17
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
jmespath==0.9.3
kombu==4.0.2
lxml==3.8.0
Markdown==2.6.8
mccabe==0.5.3
mixer==5.6.6
monotonic==1.3
ndg-httpsclient==0.4.2
oauth2client==3.0.0
oauthlib==2.0.0
olefile==0.44
opbeat==3.5.2
paramiko==2.1.1
pathlib2==2.1.0
pathspec==0.5.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.1
ply==3.8
prompt-toolkit==1.0.9
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
proto-google-cloud-pubsub-v1==0.15.4
proto-google-cloud-spanner-admin-database-v1==0.15.3
proto-google-cloud-spanner-admin-instance-v1==0.15.3
proto-google-cloud-spanner-v1==0.15.3
protobuf==3.3.0
psycopg2==2.7
ptyprocess==0.5.1
pusher==1.6
py==1.4.32
pyasn1==0.3.2
pyasn1-modules==0.0.11
pycodestyle==2.2.0
pycparser==2.17
pydub==0.20.0
pyflakes==1.3.0
Pygments==2.1.3
PyJWT==1.5.2
pyOpenSSL==17.0.0
pyowm==2.6.0
pytest==3.0.5
pytest-cov==2.4.0
pytest-django==3.1.2
python-dateutil==2.6.1
python-openid==2.2.5
pytz==2017.2
PyYAML==3.12
raven==5.32.0
redis==2.10.5
requests==2.18.4
requests-oauthlib==0.7.0
rsa==3.4.2
s3transfer==0.1.10
semantic-version==2.5.0
simplegeneric==0.8.1
six==1.10.0
sqlparse==0.2.2
tabulate==0.7.5
tenacity==4.4.0
termcolor==1.1.0
traitlets==4.3.1
twilio==5.6.0
Unidecode==0.4.19
urllib3==1.22
vine==1.1.3
wcwidth==0.1.7
websocket-client==0.44.0
whitenoise==3.3.0
